+++ libraries/transformer/src/androidTest/java/androidx/media3/transformer/ParameterizedItemExportAndroidTest.java
@@ -21,7 +21,6 @@ import static androidx.media3.test.utils.AssetInfo.JPG_ASSET;
 import static androidx.media3.test.utils.AssetInfo.MP4_ASSET;
 import static androidx.media3.test.utils.AssetInfo.PNG_ASSET;
 import static androidx.media3.test.utils.FormatSupportAssumptions.assumeFormatsSupported;
-import static androidx.media3.transformer.AndroidTestUtil.getMuxerFactoryBasedOnApi;
 import static com.google.common.truth.Truth.assertThat;
 
 import android.content.Context;
@@ -121,7 +120,6 @@ public final class ParameterizedItemExportAndroidTest {
     assumeFormatsSupported(context, testId, itemConfig.format, itemConfig.outputFormat);
     Transformer transformer =
         new Transformer.Builder(context)
-            .setMuxerFactory(getMuxerFactoryBasedOnApi())
             .setEncoderFactory(
                 new DefaultEncoderFactory.Builder(context).setEnableFallback(false).build())
             .build();